Use unsignedBigInteger for tender_id columns
Change tender_id columns from unsignedInteger to unsignedBigInteger in four migration files to match big-incrementing primary keys and avoid potential ID overflow/incompatibility. Updated migrations: 2027_04_07_000001_create_perakuan_jabatan_kertas_taklimat_tables.php, 2027_04_07_000002_create_perakuan_jabatan_pengesyoran_pembekal_table.php, 2027_04_07_000003_create_jawatankuasa_perolehan_kertas_keputusans_table.php, and 2027_04_13_000001_create_jawatankuasa_perolehan_pemilihan_tables.php. Unique constraints and other column definitions are preserved.

+++ database/migrations/2027_04_13_000001_create_jawatankuasa_perolehan_pemilihan_tables.php
@@ -11,7 +11,7 @@ return new class extends Migration
         if (!Schema::hasTable('jawatankuasa_perolehan_pemilihan_headers')) {
             Schema::create('jawatankuasa_perolehan_pemilihan_headers', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('tender_id')->unique();
+                $table->unsignedBigInteger('tender_id')->unique();
                 $table->string('keputusan_mesyuarat', 255)->nullable();
                 $table->string('kaedah_memuktamadkan_pembekal', 255)->nullable();
                 $table->string('pemilihan_berdasarkan', 255)->nullable();
